implement bandwidth scale ("extended algorithm")

+++ pypads/args.py
@@ -34,7 +34,7 @@ arg_parser.add_argument('--wavetable-size', type=int, default=262144, help='Leng
 arg_parser.add_argument('--fundamental-hz', type=int, default=440, help='Frequency of fundamental (e.g. 440 for A4)')
 arg_parser.add_argument('--bandwidth', type=int, default=50, help='Bandwidth of the first harmonic in cents')
 arg_parser.add_argument('--harmonics', type=int, default=32, help='Number of harmonics')
-arg_parser.add_argument('--scale', type=int, default=1, help='Bandwidth scale with harmonic frequency')
+arg_parser.add_argument('--scale', type=float, default=1.0, help='Bandwidth scale with harmonic frequency')
 arg_parser.add_argument('--amplitudes', required=True, type=float, nargs='+', help='Amplitude of each harmonic (zero assumed for harmonics not given)')
 
 args = vars(arg_parser.parse_args(argv[1:]))
